refactor(workflow-tools): reorganize to one file per tool with co-located schemas
- Create individual tool files in tools/ directory (11 files)
- Co-locate input schemas with their tool implementations
- Add shared types file for dependencies and context
- Simplify workspace service to aggregate tool factories
- Remove centralized schemas directory
